Mms: DUT shows 9 messages in one thread
Count the sms draft message in function
SmsRecycler.deleteMessagesForThread(). And the
draft record will auto delete in a async thread.

Use to count the sms message without draft.If there
is a draft in the conversation.Don't count it and
don't delete it.Delete the sms message older than
latestDate.Can't delete the draft.

CRs-Fixed: 424606

Change-Id: Ie55091cbe9690e3d42068a5eef5c0852cf81597e